Feature: Split unit/integration tests
