+++ packages/cubejs-clickhouse-driver/test/ClickHouseDriverTest.js
@@ -1,6 +1,7 @@
 /* globals describe, before, after, it */
 const { GenericContainer } = require("testcontainers");
 const ClickHouseDriver = require('../driver/ClickHouseDriver');
+
 require('should');
 
 describe('ClickHouseDriver', () => {
@@ -20,15 +21,15 @@ describe('ClickHouseDriver', () => {
   before(async function before() {
     this.timeout(20000);
 
-    if (!process.env.TEST_CLICKHOUSE_HOST) {
-      container = await new GenericContainer("yandex/clickhouse-server")
-        .withExposedPorts(8123)
-        .start();
-    }
+    const version = process.env.TEST_CLICKHOUSE_VERSION || 'latest';
+
+    container = await new GenericContainer('yandex/clickhouse-server', version)
+      .withExposedPorts(8123)
+      .start();
 
     config = {
       host: 'localhost',
-      port: process.env.TEST_CLICKHOUSE_HOST ? 8123 : container.getMappedPort(8123),
+      port: container.getMappedPort(8123),
     };
   });
 
@@ -192,7 +193,7 @@ describe('ClickHouseDriver', () => {
         const values = await driver.query(`SELECT * FROM ${name}.a LEFT OUTER JOIN ${name}.b ON a.x = b.x`);
         values.should.deepEqual([
           {
-            x: '1', s: 'str1', 'b.x': '0', 'b.s': null
+            x: '1', s: 'str1', 'b.x': null, 'b.s': null
           },
           {
             x: '2', s: 'str2', 'b.x': '2', 'b.s': 'str2'
