+++ packages/cubejs-query-orchestrator/test/unit/PreAggregations.test.js
@@ -105,7 +105,7 @@ describe('PreAggregations', () => {
     }
 
     jest.resetModules();
-    const QueryCache = require('../orchestrator/QueryCache');
+    const QueryCache = require('../../orchestrator/QueryCache');
     queryCache = new QueryCache(
       "TEST",
       mockDriverFactory,
@@ -122,7 +122,7 @@ describe('PreAggregations', () => {
     let preAggregations = null;
 
     beforeEach(async () => {
-      const PreAggregations = require('../orchestrator/PreAggregations');
+      const PreAggregations = require('../../orchestrator/PreAggregations');
       preAggregations = new PreAggregations(
         "TEST",
         mockDriverFactory,
@@ -146,7 +146,7 @@ describe('PreAggregations', () => {
     let preAggregations = null;
 
     beforeEach(async () => {
-      const PreAggregations = require('../orchestrator/PreAggregations');
+      const PreAggregations = require('../../orchestrator/PreAggregations');
       preAggregations = new PreAggregations(
         "TEST",
         mockDriverFactory,
@@ -171,7 +171,7 @@ describe('PreAggregations', () => {
     let preAggregations = null;
 
     beforeEach(async () => {
-      const PreAggregations = require('../orchestrator/PreAggregations');
+      const PreAggregations = require('../../orchestrator/PreAggregations');
       preAggregations = new PreAggregations(
         "TEST",
         mockDriverReadOnlyFactory,
@@ -196,7 +196,7 @@ describe('PreAggregations', () => {
     let preAggregations = null;
 
     beforeEach(async () => {
-      const PreAggregations = require('../orchestrator/PreAggregations');
+      const PreAggregations = require('../../orchestrator/PreAggregations');
       preAggregations = new PreAggregations(
         "TEST",
         mockDriverFactory,
@@ -226,7 +226,7 @@ describe('PreAggregations', () => {
     let preAggregations = null;
 
     beforeEach(async () => {
-      const PreAggregations = require('../orchestrator/PreAggregations');
+      const PreAggregations = require('../../orchestrator/PreAggregations');
       preAggregations = new PreAggregations(
         "TEST",
         () => { throw new Error('The source database factory should never be called when externalRefresh is true, as it will trigger testConnection'); },
