+++ packages/cubejs-schema-compiler/test/integration/clickhouse/ClickHouseDataSchemaCompilerTest.js
@@ -1,17 +1,21 @@
-/* globals it, describe, after */
 /* eslint-disable quote-props */
-const CompileError = require('../compiler/CompileError');
-const PostgresQuery = require('../adapter/PostgresQuery');
-const PrepareCompiler = require('./PrepareCompiler');
-const MainPrepareCompiler = require('../compiler/PrepareCompiler');
-// eslint-disable-next-line import/no-extraneous-dependencies
+/* globals describe, before, after, it */
+const CompileError = require('../../../compiler/CompileError');
+const PrepareCompiler = require('../../unit/PrepareCompiler');
+const MainPrepareCompiler = require('../../../compiler/PrepareCompiler');
 require('should');
 
 const { prepareCompiler } = PrepareCompiler;
-const dbRunner = require('./DbRunner');
+const dbRunner = require('./ClickHouseDbRunner');
 
-describe('DataSchemaCompiler', function test() {
+const { debugLog, logSqlAndParams } = require('../../unit/TestUtil');
+
+describe('ClickHouse DataSchemaCompiler', function test() {
   this.timeout(200000);
+
+  before(async function before() {
+    this.timeout(200000);
+  });
 
   after(async () => {
     await dbRunner.tearDown();
@@ -103,117 +107,6 @@ describe('DataSchemaCompiler', function test() {
     });
   });
 
-  describe('Test duplicate properties', () => {
-    const invalidSchema = `
-      cube('visitors', {
-        sql: 'select * from visitors',
-        measures: {
-          count: {
-            type: 'count',
-            sql: 'id'
-          },
-          count: {
-            type: 'count',
-            sql: 'id'
-          }
-        },
-        dimensions: {
-          date: {
-            type: 'string',
-            sql: 'date'
-          }
-        }
-      })
-    `;
-
-    const validSchema = `
-      cube('visitors', {
-        sql: 'select * from visitors',
-        measures: {
-          count: {
-            type: 'count',
-            sql: 'id'
-          }
-        },
-        dimensions: {
-          date: {
-            type: 'string',
-            sql: 'date'
-          }
-        }
-      })
-    `;
-
-    it('Should compile without error, allowJsDuplicatePropsInSchema = false, valid schema', () => {
-      const { compiler } = prepareCompiler(validSchema, { allowJsDuplicatePropsInSchema: false });
-      return compiler.compile().then(() => {
-        compiler.throwIfAnyErrors();
-      });
-    });
-
-    it('Should throw error, allowJsDuplicatePropsInSchema = false, invalid schema', () => {
-      const { compiler } = prepareCompiler(invalidSchema, { allowJsDuplicatePropsInSchema: false });
-      return compiler.compile().then(() => {
-        compiler.throwIfAnyErrors();
-        throw new Error();
-      }).catch((error) => {
-        error.should.be.instanceof(CompileError);
-        error.message.should.be.match(/Duplicate property parsing count/);
-      });
-    });
-
-    it('Should compile without error, allowJsDuplicatePropsInSchema = true, invalid schema', () => {
-      const { compiler } = prepareCompiler(invalidSchema, { allowJsDuplicatePropsInSchema: true });
-      return compiler.compile().then(() => {
-        compiler.throwIfAnyErrors();
-      });
-    });
-
-    describe('Test perfomance', () => {
-      const schema = `
-        cube('visitors', {
-          sql: 'select * from visitors',
-          measures: {
-            count: {
-              type: 'count',
-              sql: 'id'
-            },
-            duration: {
-              type: 'avg',
-              sql: 'duration'
-            },
-          },
-          dimensions: {
-            date: {
-              type: 'string',
-              sql: 'date'
-            },
-            browser: {
-              type: 'string',
-              sql: 'browser'
-            }
-          }
-        })
-      `;
-
-      it('Should compile 200 schemas in less than 2500ms * 10', async () => {
-        const repeats = 200;
-
-        const compilerWith = prepareCompiler(schema, { allowJsDuplicatePropsInSchema: false });
-        const start = new Date().getTime();
-        for (let i = 0; i < repeats; i++) {
-          delete compilerWith.compiler.compilePromise; // Reset compile result
-          await compilerWith.compiler.compile();
-        }
-        const end = new Date().getTime();
-        const time = end - start;
-        
-        time.should.be.below(2500 * 10);
-      });
-    });
-  });
-  
-
   it('calculated metrics', () => {
     const { compiler, transformer, cubeEvaluator, joinGraph } = prepareCompiler(`
     cube('visitors', {
@@ -253,7 +146,7 @@ describe('DataSchemaCompiler', function test() {
     })
     `);
     const result = compiler.compile().then(() => {
-      const query = new PostgresQuery({ joinGraph, cubeEvaluator, compiler }, {
+      const query = dbRunner.newQuery({ joinGraph, cubeEvaluator, compiler }, {
         measures: ['visitors.visitor_count'],
         timeDimensions: [{
           dimension: 'visitors.created_at',
@@ -271,14 +164,15 @@ describe('DataSchemaCompiler', function test() {
         timezone: 'America/Los_Angeles'
       });
 
-      console.log(query.buildSqlAndParams());
+      logSqlAndParams(query);
+
       return dbRunner.testQuery(query.buildSqlAndParams()).then(res => {
         res.should.be.deepEqual(
           [
-            { 'visitors__created_at_day': '2017-01-02T00:00:00.000Z', 'visitors__visitor_count': '1' },
-            { 'visitors__created_at_day': '2017-01-04T00:00:00.000Z', 'visitors__visitor_count': '1' },
-            { 'visitors__created_at_day': '2017-01-05T00:00:00.000Z', 'visitors__visitor_count': '1' },
-            { 'visitors__created_at_day': '2017-01-06T00:00:00.000Z', 'visitors__visitor_count': '2' }
+            { 'visitors__created_at_day': '2017-01-02T00:00:00.000', 'visitors__visitor_count': '1' },
+            { 'visitors__created_at_day': '2017-01-04T00:00:00.000', 'visitors__visitor_count': '1' },
+            { 'visitors__created_at_day': '2017-01-05T00:00:00.000', 'visitors__visitor_count': '1' },
+            { 'visitors__created_at_day': '2017-01-06T00:00:00.000', 'visitors__visitor_count': '2' }
           ]
         );
       });
@@ -323,7 +217,7 @@ describe('DataSchemaCompiler', function test() {
     })
     `);
     const result = compiler.compile().then(() => {
-      const query = new PostgresQuery({ joinGraph, cubeEvaluator, compiler }, {
+      const query = dbRunner.newQuery({ joinGraph, cubeEvaluator, compiler }, {
         measures: ['visitors.visitor_count'],
         dimensions: ['visitors.status'],
         timezone: 'America/Los_Angeles',
@@ -332,7 +226,7 @@ describe('DataSchemaCompiler', function test() {
         }]
       });
 
-      console.log(query.buildSqlAndParams());
+      logSqlAndParams(query);
 
       return dbRunner.testQuery(query.buildSqlAndParams()).then(res => {
         res.should.be.deepEqual(
@@ -347,50 +241,126 @@ describe('DataSchemaCompiler', function test() {
     return result;
   });
 
-  it('filtered dates', () => {
-    const { compiler, cubeEvaluator, joinGraph } = prepareCompiler(`
+  it('dynamic dimension case', () => {
+    const { compiler, transformer, cubeEvaluator, joinGraph } = prepareCompiler(`
     cube('visitors', {
       sql: \`
       select * from visitors
       \`,
+
+      measures: {
+        visitor_count: {
+          type: 'count',
+          sql: 'id'
+        }
+      },
 
       dimensions: {
         source: {
           type: 'string',
           sql: 'source'
         },
+        latitude: {
+          type: 'string',
+          sql: 'latitude'
+        },
+        enabled_source: {
+          type: 'string',
+          case: {
+            when: [{
+              sql: \`\${CUBE}.status = 3\`,
+              label: 'three'
+            }, {
+              sql: \`\${CUBE}.status = 2\`,
+              label: {
+                sql: \`\${CUBE}.source\`
+              }
+            }],
+            else: {
+              label: {
+                sql: \`\${CUBE}.source\`
+              }
+            }
+          }
+        },
         created_at: {
           type: 'time',
           sql: 'created_at'
-        },
-        updated_at: {
-          type: 'time',
-          sql: 'updated_at'
         }
       }
     })
     `);
-    const responses = [
-      [{ 'visitors__created_at': '2017-01-03T00:00:00.000Z' }],
-      [
-        { 'visitors__created_at': '2016-09-07T00:00:00.000Z' },
-        { 'visitors__created_at': '2017-01-05T00:00:00.000Z' },
-        { 'visitors__created_at': '2017-01-06T00:00:00.000Z' },
-        { 'visitors__created_at': '2017-01-07T00:00:00.000Z' }
-      ],
-      [{ 'visitors__created_at': '2017-01-07T00:00:00.000Z' }],
-      [
-        { 'visitors__created_at': '2016-09-07T00:00:00.000Z' },
-        { 'visitors__created_at': '2017-01-03T00:00:00.000Z' },
-        { 'visitors__created_at': '2017-01-05T00:00:00.000Z' },
-        { 'visitors__created_at': '2017-01-06T00:00:00.000Z' }
-      ],
-      [{ 'visitors__created_at': '2017-01-07T00:00:00.000Z' }]
-    ];
     const result = compiler.compile().then(() => {
-      const queries = ['in_date_range', 'not_in_date_range', 'on_the_date', 'before_date', 'after_date'].map((operator, index) => {
-        const filterValues = index < 2 ? ['2017-01-01', '2017-01-03'] : ['2017-01-06', '2017-01-06'];
-        return new PostgresQuery({ joinGraph, cubeEvaluator, compiler }, {
+      const query = dbRunner.newQuery({ joinGraph, cubeEvaluator, compiler }, {
+        measures: ['visitors.visitor_count'],
+        dimensions: ['visitors.enabled_source'],
+        timezone: 'America/Los_Angeles',
+        order: [{
+          id: 'visitors.enabled_source'
+        }]
+      });
+      logSqlAndParams(query);
+
+      return dbRunner.testQuery(query.buildSqlAndParams()).then(res => {
+        res.should.be.deepEqual(
+          [
+            { 'visitors__enabled_source': 'google', 'visitors__visitor_count': '1' },
+            { 'visitors__enabled_source': 'some', 'visitors__visitor_count': '2' },
+            { 'visitors__enabled_source': null, 'visitors__visitor_count': '3' },
+          ]
+        );
+      });
+    });
+
+    return result;
+  });
+
+  {
+    const { compiler, cubeEvaluator, joinGraph } = prepareCompiler(`
+      cube('visitors', {
+        sql: \`
+        select * from visitors
+        \`,
+
+        dimensions: {
+          source: {
+            type: 'string',
+            sql: 'source'
+          },
+          created_at: {
+            type: 'time',
+            sql: 'created_at'
+          },
+          updated_at: {
+            type: 'time',
+            sql: 'updated_at'
+          }
+        }
+      })
+      `);
+    const responses = [
+      [{ 'visitors__created_at': '2017-01-02T16:00:00.000' }],
+      [
+        { 'visitors__created_at': '2016-09-06T16:00:00.000' },
+        { 'visitors__created_at': '2017-01-04T16:00:00.000' },
+        { 'visitors__created_at': '2017-01-05T16:00:00.000' },
+        { 'visitors__created_at': '2017-01-06T16:00:00.000' }
+      ],
+      [{ 'visitors__created_at': '2017-01-06T16:00:00.000' }],
+      [
+        { 'visitors__created_at': '2016-09-06T16:00:00.000' },
+        { 'visitors__created_at': '2017-01-02T16:00:00.000' },
+        { 'visitors__created_at': '2017-01-04T16:00:00.000' },
+        { 'visitors__created_at': '2017-01-05T16:00:00.000' }
+      ],
+      [{ 'visitors__created_at': '2017-01-06T16:00:00.000' }]
+    ];
+    ['in_date_range', 'not_in_date_range', 'on_the_date', 'before_date', 'after_date'].map((operator, index) => {
+      const filterValues = index < 2 ? ['2017-01-01', '2017-01-03'] : ['2017-01-06', '2017-01-06'];
+      it(`filtered dates ${operator}`, async () => {
+        await compiler.compile();
+
+        const query = dbRunner.newQuery({ joinGraph, cubeEvaluator, compiler }, {
           measures: [],
           dimensions: ['visitors.created_at'],
           timeDimensions: [],
@@ -405,18 +375,14 @@ describe('DataSchemaCompiler', function test() {
           }],
           timezone: 'America/Los_Angeles'
         });
-      });
-
-      return Promise.all(queries.map(async (query, index) => {
-        console.log(query.buildSqlAndParams());
+        logSqlAndParams(query);
         const res = await dbRunner.testQuery(query.buildSqlAndParams());
 
         res.should.be.deepEqual(responses[index]);
-      }));
+      });
+      return true;
     });
-
-    return result;
-  });
+  }
 
   it('export import', () => {
     const { compiler, cubeEvaluator, joinGraph } = MainPrepareCompiler.prepareCompiler({
@@ -442,16 +408,16 @@ describe('DataSchemaCompiler', function test() {
           `
         }
       ])
-    }, { adapter: 'postgres' });
+    }, { adapter: dbRunner.adapter });
     return compiler.compile().then(() => {
-      const query = new PostgresQuery({ joinGraph, cubeEvaluator, compiler }, {
+      const query = dbRunner.newQuery({ joinGraph, cubeEvaluator, compiler }, {
         measures: ['Main.count'],
         dimensions: [],
         timeDimensions: [],
         order: [],
         timezone: 'America/Los_Angeles'
       });
-      console.log(query.buildSqlAndParams());
+      logSqlAndParams(query);
       query.buildSqlAndParams()[0].should.match(/bar/);
     });
   });
