+++ packages/cubejs-schema-compiler/test/integration/postgres/DbRunner.js
@@ -1,0 +1,3 @@
+const PostgresDBRunner = require('./PostgresDBRunner');
+
+module.exports = PostgresDBRunner;
