+++ packages/cubejs-schema-compiler/test/unit/ScaffoldingSchemaTest.js
@@ -1,7 +1,7 @@
 /* eslint-disable quote-props */
 /* globals it,describe */
-const ScaffoldingSchema = require('../scaffolding/ScaffoldingSchema');
-const ScaffoldingTemplate = require('../scaffolding/ScaffoldingTemplate');
+const ScaffoldingSchema = require('../../scaffolding/ScaffoldingSchema');
+const ScaffoldingTemplate = require('../../scaffolding/ScaffoldingTemplate');
 require('should');
 
 const driver = {
